// src/app/home/models/users.model.ts
export class User {

    _id: string;
    nom: string;
    prenom: string;
    email: string;
    password: string;
    telephone: string;
    civilite: string;
    pays_adresse: string;
    ville_adresse: string;
    rue_adresse: string;
    num_adresse: number;
    postal_adresse: number;
    userRole: string;
    idService: string

    
  }


  